refactor: extract message handler to src/content/messageHandler.ts
- Move messageListener and contentConfigUpdateListener from content.ts
  into a new content/messageHandler.ts module
- Use dependency injection (MessageHandlerDeps interface) for all
  callbacks that live in content.ts
- Export initMessageHandler() to wire deps and setupMessageListener()
  to register the chrome.runtime.onMessage listener and config sync
- Remove unused imports (Message, MessageResponse, StorageChangesObject,
  importTimes) from content.ts
